backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
boto3>=1.34
botocore>=1.34
python-dotenv>=1.0
pydantic>=2.5
mangum>=0.17.0

